Fix files name.
