+++ tests/UnitTests.Databases.SqlServer.Tests/SqlServerDatabaseInitializerDacPacTest.cs
@@ -1,28 +1,21 @@
 ﻿//-----------------------------------------------------------------------
-// <copyright file="SqlServerDatabaseInitializerDbContext.cs" company="P.O.S Informatique">
+// <copyright file="SqlServerDatabaseInitializerDacPacTest.cs" company="P.O.S Informatique">
 //     Copyright (c) P.O.S Informatique. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
 
 namespace PosInformatique.UnitTests.Databases.SqlServer.Tests
 {
-    using Microsoft.EntityFrameworkCore;
-
     [Collection("PosInformatique.UnitTests.Databases.SqlServer.Tests")]
-    public class SqlServerDatabaseInitializerDbContext : IClassFixture<SqlServerDatabaseInitializer>
+    public class SqlServerDatabaseInitializerDacPacTest : IClassFixture<SqlServerDatabaseInitializer>
     {
-        private const string ConnectionString = $"Data Source=(localDB)\\posinfo-unit-tests; Initial Catalog={nameof(SqlServerDatabaseInitializerDbContext)}; Integrated Security=True";
+        private const string ConnectionString = $"Data Source=(localDB)\\posinfo-unit-tests; Initial Catalog={nameof(SqlServerDatabaseInitializerDacPacTest)}; Integrated Security=True";
 
         private readonly SqlServerDatabase database;
 
-        public SqlServerDatabaseInitializerDbContext(SqlServerDatabaseInitializer initializer)
+        public SqlServerDatabaseInitializerDacPacTest(SqlServerDatabaseInitializer initializer)
         {
-            var optionsBuilder = new DbContextOptionsBuilder<DbContextTest>()
-                .UseSqlServer(ConnectionString);
-
-            using var context = new DbContextTest(optionsBuilder.Options);
-
-            this.database = initializer.Initialize(context);
+            this.database = initializer.Initialize("UnitTests.Databases.SqlServer.Tests.DacPac.dacpac", ConnectionString);
 
             var table = this.database.ExecuteQuery("SELECT * FROM MyTable");
 
@@ -72,37 +65,6 @@ namespace PosInformatique.UnitTests.Databases.SqlServer.Tests
 
             // Insert a row which should not be use in other tests.
             this.database.InsertInto("MyTable", new { Id = 99, Name = "Should not be here for the next unit test" });
-        }
-
-        private sealed class DbContextTest : DbContext
-        {
-            public DbContextTest(DbContextOptions<DbContextTest> options)
-                : base(options)
-            {
-            }
-
-            protected override void OnModelCreating(ModelBuilder modelBuilder)
-            {
-                modelBuilder.Entity<Table>(t =>
-                {
-                    t.ToTable("MyTable");
-
-                    t.Property(t => t.Id)
-                        .ValueGeneratedNever();
-
-                    t.Property(t => t.Name)
-                        .IsRequired()
-                        .HasMaxLength(50)
-                        .IsUnicode(false);
-                });
-            }
-        }
-
-        private sealed class Table
-        {
-            public int Id { get; set; }
-
-            public string Name { get; set; }
         }
     }
 }
