+++ tests/UnitTests.Databases.SqlServer.Tests/SqlServerDatabaseInitializerDbContextTest.cs
@@ -1,21 +1,28 @@
 ﻿//-----------------------------------------------------------------------
-// <copyright file="SqlServerDatabaseInitializerDacPac.cs" company="P.O.S Informatique">
+// <copyright file="SqlServerDatabaseInitializerDbContextTest.cs" company="P.O.S Informatique">
 //     Copyright (c) P.O.S Informatique. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
 
 namespace PosInformatique.UnitTests.Databases.SqlServer.Tests
 {
+    using Microsoft.EntityFrameworkCore;
+
     [Collection("PosInformatique.UnitTests.Databases.SqlServer.Tests")]
-    public class SqlServerDatabaseInitializerDacPac : IClassFixture<SqlServerDatabaseInitializer>
+    public class SqlServerDatabaseInitializerDbContextTest : IClassFixture<SqlServerDatabaseInitializer>
     {
-        private const string ConnectionString = $"Data Source=(localDB)\\posinfo-unit-tests; Initial Catalog={nameof(SqlServerDatabaseInitializerDacPac)}; Integrated Security=True";
+        private const string ConnectionString = $"Data Source=(localDB)\\posinfo-unit-tests; Initial Catalog={nameof(SqlServerDatabaseInitializerDbContextTest)}; Integrated Security=True";
 
         private readonly SqlServerDatabase database;
 
-        public SqlServerDatabaseInitializerDacPac(SqlServerDatabaseInitializer initializer)
+        public SqlServerDatabaseInitializerDbContextTest(SqlServerDatabaseInitializer initializer)
         {
-            this.database = initializer.Initialize("UnitTests.Databases.SqlServer.Tests.DacPac.dacpac", ConnectionString);
+            var optionsBuilder = new DbContextOptionsBuilder<DbContextTest>()
+                .UseSqlServer(ConnectionString);
+
+            using var context = new DbContextTest(optionsBuilder.Options);
+
+            this.database = initializer.Initialize(context);
 
             var table = this.database.ExecuteQuery("SELECT * FROM MyTable");
 
@@ -65,6 +72,37 @@ namespace PosInformatique.UnitTests.Databases.SqlServer.Tests
 
             // Insert a row which should not be use in other tests.
             this.database.InsertInto("MyTable", new { Id = 99, Name = "Should not be here for the next unit test" });
+        }
+
+        private sealed class DbContextTest : DbContext
+        {
+            public DbContextTest(DbContextOptions<DbContextTest> options)
+                : base(options)
+            {
+            }
+
+            protected override void OnModelCreating(ModelBuilder modelBuilder)
+            {
+                modelBuilder.Entity<Table>(t =>
+                {
+                    t.ToTable("MyTable");
+
+                    t.Property(t => t.Id)
+                        .ValueGeneratedNever();
+
+                    t.Property(t => t.Name)
+                        .IsRequired()
+                        .HasMaxLength(50)
+                        .IsUnicode(false);
+                });
+            }
+        }
+
+        private sealed class Table
+        {
+            public int Id { get; set; }
+
+            public string Name { get; set; }
         }
     }
 }
